+++ src/app/components/SettingsMenu/index.tsx
@@ -87,7 +87,7 @@ import {localize} from "../../services/utilities/localize";
 import './style.scss';
 import 'react-image-crop/dist/ReactCrop.css';
 
-export const C_VERSION = '0.27.6';
+export const C_VERSION = '0.27.7';
 export const C_CUSTOM_BG_ID = 'river_custom_bg';
 
 export const languageList = [{